refactor: removing node interface and resolver from modules
